Replace time-based IDs with raft message index
Previously, we took time.Now().UnixNano() as the ID, with some code to
ensure the values were indeed monotonically increasing. While the
design obviously worked well enough in practice, it was not as sound
as it could have been.

With this commit, we switch to the raft message index (which starts at
1 and increases from there), offset by
-message_offset=4648398125000000000 so that for upgraded existing
RobustIRC networks, the message IDs remain strictly monotonically
increasing.

In case this commit causes any issues and you need to downgrade a
RobustIRC network, increase -message_offset by more time than the
commit has been running. E.g., if the commit had been running for a
day, increase -message_offset by 86000000000000.

fixes issue #150
fixes issue #123

This commit increases the network throughput from 2000 messages/s to
3000 messages/s:

https://snapshot.raintank.io/dashboard/snapshot/d6U0rDZnQ3RY0cdUk13zU5FXfTFT3N2z

+++ internal/outputstream/serialization.go
@@ -56,19 +56,19 @@ func unmarshalMessageBatch(buffer []byte) *messageBatch {
 	n += 8
 	for i := 0; i < len(result.Messages); i++ {
 		msg := &(result.Messages[i])
-		msg.Id.Id = int64(binary.LittleEndian.Uint64(buffer[n:]))
+		msg.Id.Id = binary.LittleEndian.Uint64(buffer[n:])
 		n += 8
-		msg.Id.Reply = int64(binary.LittleEndian.Uint64(buffer[n:]))
+		msg.Id.Reply = binary.LittleEndian.Uint64(buffer[n:])
 		n += 8
 		lenData = int(binary.LittleEndian.Uint64(buffer[n:]))
 		n += 8
 		msg.Data = string(buffer[n : n+lenData])
 		n += lenData
 		lenInterestingFor = binary.LittleEndian.Uint64(buffer[n:])
-		msg.InterestingFor = make(map[int64]bool, lenInterestingFor)
+		msg.InterestingFor = make(map[uint64]bool, lenInterestingFor)
 		n += 8
 		for j := uint64(0); j < lenInterestingFor; j++ {
-			msg.InterestingFor[int64(binary.LittleEndian.Uint64(buffer[n:]))] = true
+			msg.InterestingFor[uint64(binary.LittleEndian.Uint64(buffer[n:]))] = true
 			n += 8
 		}
 	}
